refactor: 월 변경 컴포넌트 수정, constance 폴더 파일 대소문자 수정
